Fixed errors and indentation
I fixed some errors, to list some, originally before this edit if the javascript was unable to find any bots at all, say from a possible connection problem then the `0 Results, Page 1 of 1` footer would display `undefined Page NaN of 1` now it'll display `0 Results, Page 1 of 1` instead, and again if it was unable to find any bots it would spit out an error for cannot read property length of undefined, I added an if check before that code block so it won't execute if there are no bot's found. I have also fixed the indentation of the code to make it prettier to look at and added semi-colons to the end of the lines in the javascript because the first half already had semi-colons. I did this to keep it the same.